Moved ticks & playhead to timeline
Since these components are only used in the timeline, they were moved to their folder from the top-level hierarchy.
