+++ src/components/timeline/playhead.tsx
