+++ src/components/timeline/ticks.tsx
